Refactor tests and fixtures to organise them by platform
This factoring breaks down the tests and fixtures by platform so that
they are easier to locate than when they're all in one giant conftest
file.

It adds some fixtures for PacBio that are for future tests (marked as
TODO).
