+++ src/npg_irods/db/mlwh.py
@@ -19,10 +19,28 @@
 
 """Business logic API and schema-level API for the ML warehouse."""
 
+import enum
 from typing import Type
 
-from sqlalchemy import DateTime, ForeignKey, Integer, String, Text
+from sqlalchemy import (
+    BigInteger,
+    DateTime,
+    Enum,
+    ForeignKey,
+    Index,
+    Integer,
+    String,
+    Text,
+)
 from sqlalchemy.orm import DeclarativeBase, Mapped, Session, mapped_column, relationship
+
+
+class Platform(enum.Enum):
+    """Sequencing platform values for SeqProductIrodsLocations.seq_platform_name"""
+
+    Illumina = 1
+    ONT = 2
+    PacBio = 3
 
 
 class Base(DeclarativeBase):
@@ -32,7 +50,7 @@ class Base(DeclarativeBase):
 class Sample(Base):
     __tablename__ = "sample"
 
-    id_sample_tmp = mapped_column(Integer, primary_key=True)
+    id_sample_tmp = mapped_column(Integer, primary_key=True, autoincrement=True)
     id_lims = mapped_column(String(10), nullable=False)
     id_sample_lims = mapped_column(String(20), nullable=False)
     last_updated = mapped_column(DateTime, nullable=False)
@@ -58,6 +76,10 @@ class Sample(Base):
         "OseqFlowcell", back_populates="sample"
     )
 
+    pac_bio_run: Mapped["PacBioRun"] = relationship(
+        "PacBioRun", back_populates="sample"
+    )
+
     def __repr__(self):
         return (
             f"<Sample pk={self.id_sample_tmp} id_sample_lims={self.id_sample_lims} "
@@ -68,7 +90,7 @@ class Sample(Base):
 class Study(Base):
     __tablename__ = "study"
 
-    id_study_tmp = mapped_column(Integer, primary_key=True)
+    id_study_tmp = mapped_column(Integer, primary_key=True, autoincrement=True)
     id_lims = mapped_column(String(10), nullable=False)
     id_study_lims = mapped_column(String(20), nullable=False)
     last_updated = mapped_column(DateTime, nullable=False)
@@ -94,6 +116,8 @@ class Study(Base):
         "OseqFlowcell", back_populates="study"
     )
 
+    pac_bio_run: Mapped["PacBioRun"] = relationship("PacBioRun", back_populates="study")
+
     def __repr__(self):
         return f"<Study pk={self.id_study_tmp} id_study_lims={self.id_study_lims}>"
 
@@ -101,7 +125,7 @@ class Study(Base):
 class IseqFlowcell(Base):
     __tablename__ = "iseq_flowcell"
 
-    id_iseq_flowcell_tmp = mapped_column(Integer, primary_key=True)
+    id_iseq_flowcell_tmp = mapped_column(Integer, primary_key=True, autoincrement=True)
     last_updated = mapped_column(DateTime, nullable=False)
     recorded_at = mapped_column(DateTime, nullable=False)
     id_sample_tmp = mapped_column(
@@ -133,7 +157,9 @@ class IseqFlowcell(Base):
 class IseqProductMetrics(Base):
     __tablename__ = "iseq_product_metrics"
 
-    id_iseq_pr_metrics_tmp = mapped_column(Integer, primary_key=True)
+    id_iseq_pr_metrics_tmp = mapped_column(
+        Integer, primary_key=True, autoincrement=True
+    )
     id_iseq_product = mapped_column(String(64), nullable=False, unique=True)
     last_changed = mapped_column(DateTime)
     id_iseq_flowcell_tmp = mapped_column(
@@ -161,7 +187,7 @@ class IseqProductMetrics(Base):
 class OseqFlowcell(Base):
     __tablename__ = "oseq_flowcell"
 
-    id_oseq_flowcell_tmp = mapped_column(Integer, primary_key=True)
+    id_oseq_flowcell_tmp = mapped_column(Integer, primary_key=True, autoincrement=True)
     id_flowcell_lims = mapped_column(String(255), nullable=False)
     last_updated = mapped_column(DateTime, nullable=False)
     recorded_at = mapped_column(DateTime, nullable=False)
@@ -197,6 +223,128 @@ class OseqFlowcell(Base):
             f"slot={self.instrument_slot} "
             f"flowcell={self.flowcell_id}>"
         )
+
+
+class PacBioRun(Base):
+    __tablename__ = "pac_bio_run"
+
+    id_pac_bio_tmp = mapped_column(Integer, primary_key=True, autoincrement=True)
+    last_updated = mapped_column(DateTime, nullable=False)
+    recorded_at = mapped_column(DateTime, nullable=False)
+    id_sample_tmp = mapped_column(
+        Integer, ForeignKey("sample.id_sample_tmp"), nullable=False
+    )
+    id_study_tmp = mapped_column(
+        Integer, ForeignKey("study.id_study_tmp"), nullable=False
+    )
+    pac_bio_run_name = mapped_column(String(255))
+    well_label = mapped_column(String(255), nullable=False)
+    plate_number = mapped_column(Integer)
+    id_lims = mapped_column(String(10), nullable=False)
+    id_pac_bio_run_lims = mapped_column(String(20), nullable=False)
+    tag_identifier = mapped_column(String(30))
+    tag_sequence = mapped_column(String(30))
+    tag_set_id_lims = mapped_column(String(20))
+    tag_set_name = mapped_column(String(100))
+    tag2_sequence = mapped_column(String(30))
+    tag2_set_id_lims = mapped_column(String(20))
+    tag2_set_name = mapped_column(String(100))
+    tag2_identifier = mapped_column(String(30))
+
+    sample: Mapped["Sample"] = relationship("Sample", back_populates="pac_bio_run")
+    study: Mapped["Study"] = relationship("Study", back_populates="pac_bio_run")
+    pac_bio_product_metrics: Mapped["PacBioProductMetrics"] = relationship(
+        "PacBioProductMetrics", back_populates="pac_bio_run"
+    )
+
+    def __repr__(self):
+        return (
+            f"<PacBioRun run_name={self.pac_bio_run_name} "
+            f"well_label={self.well_label} "
+            f"plate_number={self.plate_number}>"
+        )
+
+
+class PacBioRunWellMetrics(Base):
+    __tablename__ = "pac_bio_run_well_metrics"
+    __table_args__ = (
+        Index(
+            "pac_bio_metrics_run_well",
+            "pac_bio_run_name",
+            "well_label",
+            "plate_number",
+            unique=True,
+        ),
+    )
+
+    id_pac_bio_rw_metrics_tmp = mapped_column(
+        Integer, primary_key=True, autoincrement=True
+    )
+
+    last_changed = mapped_column(DateTime)
+    id_pac_bio_product = mapped_column(String(64), nullable=False)
+    pac_bio_run_name = mapped_column(String(255), nullable=False)
+    well_label = mapped_column(String(255), nullable=False)
+    plate_number = mapped_column(Integer)
+
+    pac_bio_product_metrics: Mapped["PacBioProductMetrics"] = relationship(
+        "PacBioProductMetrics", back_populates="pac_bio_run_well_metrics"
+    )
+
+    def __repr__(self):
+        return (
+            f"<PacBioRunWellMetrics run_name={self.pac_bio_run_name} "
+            f"well_label={self.well_label} "
+            f"plate_number={self.plate_number} "
+            f"product={self.id_pac_bio_product}>"
+        )
+
+
+class PacBioProductMetrics(Base):
+    __tablename__ = "pac_bio_product_metrics"
+    __table_args__ = (
+        Index(
+            "pac_bio_metrics_product",
+            "id_pac_bio_tmp",
+            "id_pac_bio_rw_metrics_tmp",
+            unique=True,
+        ),
+    )
+
+    last_changed = mapped_column(DateTime)
+    id_pac_bio_pr_metrics_tmp = mapped_column(Integer, primary_key=True)
+    id_pac_bio_rw_metrics_tmp = mapped_column(
+        ForeignKey("pac_bio_run_well_metrics.id_pac_bio_rw_metrics_tmp"),
+        nullable=False,
+        index=True,
+    )
+    id_pac_bio_tmp = mapped_column(
+        ForeignKey("pac_bio_run.id_pac_bio_tmp"), nullable=False, index=True
+    )
+    id_pac_bio_product = mapped_column(String(64), nullable=False, unique=True)
+
+    pac_bio_run_well_metrics: Mapped["PacBioRunWellMetrics"] = relationship(
+        "PacBioRunWellMetrics", back_populates="pac_bio_product_metrics"
+    )
+    pac_bio_run: Mapped["PacBioRun"] = relationship(
+        "PacBioRun", back_populates="pac_bio_product_metrics"
+    )
+
+
+class SeqProductIrodsLocations(Base):
+    __tablename__ = "seq_product_irods_locations"
+
+    id_seq_product_irods_locations_tmp = mapped_column(
+        BigInteger, primary_key=True, autoincrement=True
+    )
+    created = mapped_column(DateTime)
+    last_changed = mapped_column(DateTime)
+    id_product = mapped_column(String(64), nullable=False)
+    seq_platform_name = mapped_column(Enum(Platform), nullable=False)
+    pipeline_name = mapped_column(String(32), nullable=False)
+    irods_root_collection = mapped_column(String(255), nullable=False)
+    irods_data_relative_path = mapped_column(String(255))
+    irods_secondary_data_relative_path = mapped_column(String(255))
 
 
 def find_consent_withdrawn_samples(sess: Session) -> list[Type[Sample]]:
