+++ tests/data/pacbio/synthetic/r12345_20246789_98765/1_A01/m12345_246789_987655_s3.fail_reads.bc1000.bam
@@ -1,0 +1,1 @@
+79bf05ca-b91b-11ee-8a6e-93dd9eb1bda2
