+++ tests/data/pacbio/synthetic/r12345_20246789_98765/1_A01/m12345_246789_987655_s3.fail_reads.bc1000.bam.pbi
@@ -1,0 +1,1 @@
+79c0b4c4-b91b-11ee-9357-e700022c250d
