+++ tests/data/pacbio/synthetic/r12345_20246789_98765/1_A01/m12345_246789_987655_s3.fail_reads.bc1000.consensusreadset.xml
@@ -1,0 +1,1 @@
+79c28d6c-b91b-11ee-acc6-cf26908f472f
