+++ tests/data/pacbio/synthetic/r12345_20246789_98765/1_A01/m12345_246789_987655_s3.fail_reads.unassigned.bam
@@ -1,0 +1,1 @@
+79c5cc48-b91b-11ee-9b48-a7792c558017
