+++ tests/data/pacbio/synthetic/r12345_20246789_98765/1_A01/m12345_246789_987655_s3.fail_reads.unassigned.bam.pbi
@@ -1,0 +1,1 @@
+79c75cfc-b91b-11ee-aae8-d353958c74a2
