+++ tests/data/pacbio/synthetic/r12345_20246789_98765/1_A01/m12345_246789_987655_s3.fail_reads.unassigned.consensusreadset.xml
@@ -1,0 +1,1 @@
+79c8f580-b91b-11ee-9c75-2b616c425185
