+++ tests/data/pacbio/synthetic/r12345_20246789_98765/1_A01/m12345_246789_987655_s3.hifi_reads.bc1000.bam
@@ -1,0 +1,1 @@
+79ca919c-b91b-11ee-bd78-0f366589f40e
