+++ tests/data/pacbio/synthetic/r12345_20246789_98765/1_A01/m12345_246789_987655_s3.hifi_reads.bc1000.bam.pbi
@@ -1,0 +1,1 @@
+79cc0914-b91b-11ee-a7b1-4f4d0b768eec
