+++ tests/data/pacbio/synthetic/r12345_20246789_98765/1_A01/m12345_246789_987655_s3.hifi_reads.bc1000.consensusreadset.xml
@@ -1,0 +1,1 @@
+79cd969e-b91b-11ee-8cdd-9f58f14ba415
