+++ tests/data/pacbio/synthetic/r12345_20246789_98765/1_A01/m12345_246789_987655_s3.hifi_reads.consensusreadset.xml
@@ -1,0 +1,1 @@
+79cf1a14-b91b-11ee-ba8b-331134c71f61
