+++ tests/data/pacbio/synthetic/r12345_20246789_98765/1_A01/m12345_246789_987655_s3.hifi_reads.unassigned.bam.pbi
@@ -1,0 +1,1 @@
+79d22e66-b91b-11ee-ad97-433f477a33d5
