+++ tests/data/pacbio/synthetic/r12345_20246789_98765/1_A01/m12345_246789_987655_s3.hifi_reads.unassigned.consensusreadset.xml
@@ -1,0 +1,1 @@
+79d3b38a-b91b-11ee-9ec5-c31f412ff2e6
