+++ tests/data/pacbio/synthetic/r12345_20246789_98765/1_A01/m12345_246789_987655_s3.sts.xml
@@ -1,0 +1,1 @@
+79d6d51a-b91b-11ee-a48f-7fe30217a4b7
