+++ tests/data/pacbio/synthetic/r12345_20246789_98765/1_A01/m12345_246789_987655_s3.zmw_metrics.json.gz
@@ -1,0 +1,1 @@
+79d858a4-b91b-11ee-9c06-8fd32712288f
